Added rudimentary Ajax interaction, see description:
When clicking on part of legislation, the "about" attribute value is
sent to the server. The shorter entity description used as nodes in the
network is returned, alongside the original "about" value.

The container for the extra information "floats" so always stays on the
same place while the user can simply scroll through the legislation.