Lots of additions to web page. Note that css and javascript are not (yet) under source control
Css and javascript are hosted in dropbox public folder.

Additions:
When clicking on entity, sorted related entities are shown.
View for details on related entities created, but not yet filled with
actual content from those sources.

Improvements:
Lots of css and javascript to create a nice user experience. Small and
wide screens are supported.