Metalex data for selected results is now retrieved and shown.
Work URI's aren't saved as attributes in the network anymore. Instead,
they are put in a regular dictionary which can be queried without
loading the rest of the graph data.

Created SparqlHelper class to house functions that have to interact
with the sparql endpoint and doc.metalex.eu in general.